fix: モデルのキャスト誤りを直し、パック/パス購入のテストを追加する
カバレッジを計測したところ BuyPackUseCase / BuyPassUseCase が0%で、
トランザクションを _BaseBuyUseCase に集約した変更が未検証だった。
テストを書いたところ、パス購入の効果が壊れていることが判明した。

## 見つかった不具合

MstInAppPurchaseEffect のキャストが実際のカラム型と食い違っていた。

  'effect_type' => 'integer'   // enum(varchar) なので 'ExpBoost' が '0' になる
  'value' => 'integer'         // decimal(10,2) なので 1.50 が 1 になる

この結果、パスを購入しても効果が effect_type='' / value=1 で保存され、
**購入した効果が一切適用されない**状態だった。例外は出ないため気づけない。

同種の誤りを全モデル×DBスキーマで突き合わせ、さらに2件見つけて直した。

- MstItem::value: DBは double。integerで小数が落ちていた
- LogUnit::mst_unit_id: DBは varchar。'unit_xxx' が 0 で記録されていた

## 再発防止

CastsMatchSchemaTest を追加した。全モデルの $casts を
information_schema のカラム型と突き合わせ、食い違いを検出する。
誤ったキャストは例外にならず値だけが壊れるため、静的解析では拾えない。

## テスト

BuyPackUseCaseTest / BuyPassUseCaseTest を追加（共通のフィクスチャは
PreparesPurchaseFixtures トレイトに切り出し）。
購入・VIP付与・課金ログが1トランザクションで確定することを確認している。

+++ api/app/Models/Mst/MstItem.php
@@ -9,7 +9,7 @@ namespace App\Models\Mst;
  * @property string $id
  * @property string $type
  * @property string $effect
- * @property int $value
+ * @property float $value
  */
 class MstItem extends _BaseMst
 {
@@ -34,7 +34,8 @@ class MstItem extends _BaseMst
     /** @var array<string, string> */
     protected $casts = [
         'deploy_key' => 'integer',
-        'value' => 'integer',
+        // DBは double。integerにすると小数が落ちる
+        'value' => 'float',
     ];
 
     public $timestamps = true;
@@ -58,9 +59,9 @@ class MstItem extends _BaseMst
     /**
      * アイテム値を取得
      */
-    public function getValue(): int
+    public function getValue(): float
     {
-        return $this->getAttribute('value');
+        return (float) $this->getAttribute('value');
     }
 
     /**
